refactor(tui): split content.go into per-panel files
Mechanical split of the 2176-LOC internal/tui/content.go into 13
per-panel files. ContentModel struct + top-level Update/View dispatch
stay in content.go (now 447 LOC).

Each panel concern owns its own state lazy-init, size propagation,
message handlers, and render helper:
- content_pipelines.go / content_pipelines_update.go (pipelines + compose)
- content_personas.go, content_contracts.go, content_skills.go,
  content_health.go, content_issues.go, content_prs.go,
  content_suggest.go, content_ontology.go (alternative views)
- content_navigation.go (cycleView, setView, numberKeyToView)
- content_routing.go (handleAlternativeViewEnter/Escape,
  routeToActiveList/Detail multi-panel dispatch)

Helpers introduced are package-private and follow a consistent shape:
ensureXxxModels, setSizeXxx, updateXxxMessage (returns handled bool),
viewXxx. No symbol renames, no struct changes, no behaviour change.

Closes #1505 (mechanical split). ContentProviders aggregator
replacement (Open/Closed violation) tracked separately.
